исправил стили для модального окна на бирже
