+++ src/components/Modals/IdeaMarketModal/IdeaMarketRightSide.vue
@@ -25,39 +25,45 @@ function openMarketInfo() {
 </script>
 
 <template>
-  <div class="header-buttons w-100 p-2 bg-white rounded">
-    <Button
-      :class-name="!isChangeContent ? 'btn-primary w-100' : 'btn-secondary w-100'"
-      @click="openMarketInfo"
-    >
-      Информация
-    </Button>
-    <Button
-      :class-name="isChangeContent ? 'btn-primary w-100' : 'btn-secondary w-100'"
-      @click="openMarketSkillsRadarCharts"
-    >
-      Компетенции
-    </Button>
-  </div>
+  <div class="right-side w-100">
+    <div class="right-side__header w-100 p-2 bg-white rounded">
+      <Button
+        :class-name="!isChangeContent ? 'btn-primary w-100' : 'btn-secondary w-100'"
+        @click="openMarketInfo"
+      >
+        Информация
+      </Button>
+      <Button
+        :class-name="isChangeContent ? 'btn-primary w-100' : 'btn-secondary w-100'"
+        @click="openMarketSkillsRadarCharts"
+      >
+        Компетенции
+      </Button>
+    </div>
 
-  <div v-if="isChangeContent">
-    <IdeaMarketSkills
-      :skills="skills"
-      v-model:skillsRequestTeam="skillsRequestTeam"
-      v-model:skillsAcceptedTeam="skillsAcceptedTeam"
-      :idea="idea"
+    <div v-if="isChangeContent">
+      <IdeaMarketSkills
+        :skills="skills"
+        v-model:skillsRequestTeam="skillsRequestTeam"
+        v-model:skillsAcceptedTeam="skillsAcceptedTeam"
+        :idea="idea"
+      />
+    </div>
+
+    <IdeaMarketInfo
+      v-else
+      :idea-market="idea"
+      :market="props.market"
     />
   </div>
-
-  <IdeaMarketInfo
-    v-else
-    :idea-market="idea"
-    :market="props.market"
-  />
 </template>
 
 <style lang="scss" scoped>
-.header-buttons {
-  @include flexible(flex-start, space-between, $gap: 8px);
+.right-side {
+  @include flexible(flex-start, flex-start, column, $gap: 16px);
+
+  &__header {
+    @include flexible(flex-start, space-between, $gap: 8px);
+  }
 }
 </style>
